added redux store with reducers and middleware chain
