+++ package.json
@@ -34,7 +34,10 @@
     "react-dark-mode-toggle": "0.0.5",
     "react-dom": "16.13.1",
     "react-icons": "3.10.0",
+    "react-redux": "7.2.2",
     "react-toggle": "4.1.1",
+    "redux": "4.0.5",
+    "redux-thunk": "2.3.0",
     "styled-components": "5.1.1",
     "tinycolor2": "1.4.1",
     "use-dark-mode": "2.3.1"
