+++ src/components/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import PalettePicker from './PalettePickerFn';
+import PalettePicker from './PalettePicker';
 import './App.css';
 export default function App() {
   return (
