+++ src/index.js
@@ -1,8 +1,13 @@
 import React from 'react';
 import { render } from 'react-dom';
 import App from './components/App';
+import { Provider } from 'react-redux';
+import store from './store';
 
 render(
-  <App />,
+  <Provider store={store}>
+    <App />
+  </Provider>
+  ,
   document.getElementById('root')
 );
